added authorization for PUT, GET findOne, and DELETE

+++ server/controller/todo.js
@@ -1,5 +1,5 @@
 const {Todo} = require('../models/index')
-const tokenAuth = require('../middleware/tokenAuth')
+
 
 class TodoController {
     static async getTodoList(req,res) {
@@ -13,11 +13,13 @@ class TodoController {
 
     static async createTodo(req,res) {
         // const { title,description,status,due_date } = req.body
+        console.log(req.header.UserId);
         const params = {
             title: req.body.title,
             description: req.body.description,
             status: req.body.status,
-            due_date: req.body.due_date
+            due_date: req.body.due_date,
+            UserId: req.userData.id
         }
         try {
             const createTodo = await Todo.create(params)
@@ -42,7 +44,7 @@ class TodoController {
 
     static async getTodoById(req,res) {
         try {
-            const todo = await Todo.findOne({where:{id:req.params.id}})
+            const todo = await Todo.findOne({where:{id:req.params.id, UserId: req.userData.id}})
             res.status(200).json(todo)
         } catch (err) {
             res.status(400).json(err)
@@ -52,7 +54,7 @@ class TodoController {
     static async updateTodo(req,res) {
         const { title,description,status,due_date } = req.body
         try {
-            const update = await Todo.update({ title,description,status,due_date },{where:{id:req.params.id}})
+            const update = await Todo.update({ title,description,status,due_date },{where:{id:req.params.id, UserId: req.userData.id}})
             res.status(201).json(update)
         } catch (err) {
             res.status(400).json(err)
@@ -60,7 +62,7 @@ class TodoController {
     }
     static async deleteTodo (req,res) {
         try {
-            const todo = await Todo.destroy({where:{id:req.params.id}})
+            const todo = await Todo.destroy({where:{id:req.params.id, UserId: req.userData.id}})
             res.status(200).json(todo)
         } catch (err) {
             res.status(400).json(err)
